// resources/js/game/components/actions/partials/floating-cards/crafting-section/sections/batch-crafting/enums/batch-crafting-screen-names.ts
export enum BatchCraftingScreenNames {
  INTRODUCTION = 'introduction',
  TYPE = 'type',
  MODE = 'mode',
  CRAFT_AMOUNT = 'craft-amount',
  CRAFT_EXPERIENCE = 'craft-experience',
  CRAFT_SET_OUTPUT = 'craft-set-output',
  CRAFT_SET = 'craft-set',
  CRAFT_EVENT = 'craft-event',
  CRAFT_AND_ENCHANT_MODE = 'craft-and-enchant-mode',
  CRAFT_AND_ENCHANT_AMOUNT_OUTPUT = 'craft-and-enchant-amount-output',
  CRAFT_AND_ENCHANT_AMOUNT = 'craft-and-enchant-amount',
  CRAFT_AND_ENCHANT_EXPERIENCE = 'craft-and-enchant-experience',
  CRAFT_AND_ENCHANT_SET_OUTPUT = 'craft-and-enchant-set-output',
  CRAFT_AND_ENCHANT_SET = 'craft-and-enchant-set',
  CRAFT_AND_ENCHANT_SET_ENCHANTMENTS = 'craft-and-enchant-set-enchantments',
  ENCHANT_EVENT = 'enchant-event',
  ALCHEMY_MODE = 'alchemy-mode',
  ALCHEMY_AMOUNT = 'alchemy-amount',
  ALCHEMY_EXPERIENCE = 'alchemy-experience',
  HOLY_OILS_MODE = 'holy-oils-mode',
  HOLY_OILS_SELECTED_ITEMS = 'holy-oils-selected-items',
  HOLY_OILS_SET = 'holy-oils-set',
  TRINKETRY = 'trinketry',
  RUNNING = 'running',
}